Make ThunkBuilder into a more extensible class
In preparation for creating a JitBuilder extension for ThunkBuilder,
to facilitate the client API creation, the ThunkBuilder class needs
to become more like a fully extensible with an independent header.
