+++ compiler/ilgen/CMakeLists.txt
@@ -28,6 +28,6 @@ compiler_library(ilgen
 	${CMAKE_CURRENT_LIST_DIR}/OMRBytecodeBuilder.cpp
 	${CMAKE_CURRENT_LIST_DIR}/OMRIlType.cpp
 	${CMAKE_CURRENT_LIST_DIR}/OMRTypeDictionary.cpp
-	${CMAKE_CURRENT_LIST_DIR}/ThunkBuilder.cpp
+	${CMAKE_CURRENT_LIST_DIR}/OMRThunkBuilder.cpp
 	${CMAKE_CURRENT_LIST_DIR}/VirtualMachineOperandStack.cpp
 )
